# MNE connectivity 2x speed-up