--- a/py_neuromodulation/nm_mne_connectivity.py
+++ b/py_neuromodulation/nm_mne_connectivity.py
@@ -8,7 +8,6 @@ from py_neuromodulation.nm_types import NMBaseModel
 
 if TYPE_CHECKING:
     from py_neuromodulation.nm_settings import NMSettings
-    from mne import Epochs
     from mne.io import RawArray
 
 
@@ -26,11 +25,14 @@ class MNEConnectivity(NMFeature):
     ) -> None:
         from mne import create_info
 
-        self.ch_names = ch_names
         self.settings = settings
+        
+        self.ch_names = ch_names
+        self.sfreq = sfreq
+        
+        # Params used by spectral_connectivity_epochs
         self.mode = settings.mne_connectivity.mode
         self.method = settings.mne_connectivity.method
-        self.sfreq = sfreq
 
         self.fbands = settings.frequency_ranges_hz
         self.fband_ranges: list = []
@@ -40,6 +42,8 @@ class MNEConnectivity(NMFeature):
         self.raw_array: "RawArray"
         self.events: np.ndarray
         self.prev_batch_shape: tuple = (-1, -1)  # sentinel value
+        self.prev_sfreq = self.sfreq
+        self.ch_names = ch_names
 
     def calc_feature(self, data: np.ndarray, features_compute: dict) -> dict:
         from mne.io import RawArray
@@ -47,7 +51,7 @@ class MNEConnectivity(NMFeature):
         from mne_connectivity import spectral_connectivity_epochs
 
         time_samples_s = data.shape[1] / self.sfreq
-        epoch_length: float = 1  # TODO: Make this a parameter
+        epoch_length: float = 1  # TODO: Make this a parameter?
 
         if epoch_length > time_samples_s:
             raise ValueError(
@@ -55,7 +59,9 @@ class MNEConnectivity(NMFeature):
                 f" are longer than the passed data array {np.round(time_samples_s, 2)}s"
             )
 
-        if data.shape != self.prev_batch_shape:
+        # Only reinitialize the raw_array and epochs object if the data shape has changed
+        # That could mean that the channels have been re-selected, or we're in the last batch
+        if data.shape != self.prev_batch_shape or self.prev_sfreq != self.sfreq:
             self.raw_array = RawArray(
                 data=data,
                 info=self.raw_info,
@@ -63,9 +69,7 @@ class MNEConnectivity(NMFeature):
                 verbose=False,
             )
 
-            # self.events = make_fixed_length_events(
-            #     self.raw_array, duration=epoch_length, overlap=0
-            # )
+            # self.events = make_fixed_length_events(self.raw_array, duration=epoch_length)
             # Equivalnet code for those parameters:
             self.events = np.column_stack(
                 (
@@ -94,14 +98,13 @@ class MNEConnectivity(NMFeature):
                 verbose=False,
             )
 
-            self.epochs.metadata = pd.DataFrame(
-                index=np.arange(self.events.shape[0]), columns=["column1", "column2"]
-            )
+            # Trick the function "spectral_connectivity_epochs" into not calling "add_annotations_to_metadata"
+            self.epochs._metadata = pd.DataFrame(index=np.arange(self.events.shape[0]))
 
         else:
-            # self.raw_array._data = data
             # As long as the initialization parameters, channels, sfreq and batch size are the same
             # We can re-use the existing epochs object by updating the raw data
+            self.raw_array._data = data
             self.epochs._raw = self.raw_array
 
         # n_jobs is here kept to 1, since setup of the multiprocessing Pool
@@ -134,6 +137,8 @@ class MNEConnectivity(NMFeature):
                 key = "_".join(["ch1", self.method, str(conn), fband])
                 features_compute[key] = fband_mean[conn]
 
+        # Store current experiment parameters to check if re-initialization is needed
         self.prev_batch_shape = data.shape
+        self.prev_sfreq = self.sfreq
 
         return features_compute