@timonmerk 2024-06-20
Hey @toni-neurosc, thanks for this PR. I think it would be best to discuss the PARRM and mne-connectivity optimization with @tsbinns, since he's a developer of both tools. They were probably not developed for real-time use, so I assume that there is room for speed improvement.