@tsbinns 2024-06-24
I realise this is closed, but just some comments. Can move to a new issue if preferred.

---

> @toni-neurosc I also noticed that the function spectral_connectivity_epochs was calling another function, add_annotations_to_metadata, that was doing some serious outsourcing to Pandas, which not only was super-slow, but was also doing 0 changes to the Epochs.annotations property that it was manipulating and that was also never used after that for anything. So I checked the conditional if statement that triggered the call to add_annotations_to_metadata and added a little hack that makes the condition fail so it's not called. But it's super-weird that I need to do that to get reasonable behavior from spectral_connectivity_epochs.

[`add_annotations_to_metadata()`](https://mne.tools/stable/generated/mne.Epochs.html#mne.Epochs.add_annotations_to_metadata) is a method of the `Epochs` class. It does not change the `annotations` attr, but rather adds this annotation information to the `metadata` attr. This metadata is then passed through and stored in the returned connectivity object, I imagine as a way of preserving info about the data from which connectivity was derived.

If you have annotations in your `Epochs` and you have not already added these to `metadata`, `spectral_connectivity_epochs()` will do this for you.

There are two less hacky solutions than modifying the private `Epochs._metadata` attr:
1. Set `Epochs.annotations=None` before computing connectivity (`add_annotations_to_metadata()` checks this first, and if so, does not make a new `DataFrame`)
2. Pass in the epoch data as an array to `spectral_connectivity_epochs()` and avoid this logic of checking for annotations/metadata

---

> @toni-neurosc Also, I have another question about the MNE-connectivity feature: how come the channels are hard-coded in the calc_feature function?
I'm talking about this argument to spectral_connectivity_epochs:
indices=(np.array([0, 0, 1, 1]), np.array([2, 3, 2, 3]))

Yeah, I would have thought this should also be a parameter the end-user can change, otherwise you are only computing connectivity from this specific combination of channels.

---

> @timonmerk They were probably not developed for real-time use, so I assume that there is room for speed improvement.

Perhaps there are some optimisations that could be made within MNE-Connectivity, but a lot of the overhead comes from thing's like `_assemble_spectral_params`, instantiating the connectivity object, and a bunch of other internal checks that are being re-run every time you pass in data for a new window as @toni-neurosc points out. These types of things could be omitted for a shorter run-time, but that's not gonna come from `spectral_connectivity_epochs()`, it would require a custom implementation of the pipeline.

---

When checking the diff I also noticed this:

https://github.com/neuromodulation/py_neuromodulation/blob/dc285675a3556a50fd48e03ee0f9d6fe72c35ddb/py_neuromodulation/nm_mne_connectivity.py#L87-L91

`spectral_connectivity_epochs()` should work as long as the data has shape `(n_epochs x n_channels x n_times)`, even if `n_epochs=1`. Is that not the case?